% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forcing_temperature.R
\name{get_forcing_temperature}
\alias{get_forcing_temperature}
\title{Read in temperature forcing by box or from output.nc}
\usage{
get_forcing_temperature(param.ls, plotFigs = F)
}
\arguments{
\item{param.ls}{list. List of file paths from Atlantis model}

\item{plotFigs}{logical. If TRUE, plots are generated}
}
\value{
tibble. Data frame of temperature data by box/layer/time
}
\description{
To ge a better idea of how temperature changes in space
and time

Important to discern if species distribution changes are due
to changes in temperature or something else or to identify which regions
a species might leave if temperature was increased
under climate scenarios
}
