% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_FXXX_SY.R
\name{get_param_FXXX_SY}
\alias{get_param_FXXX_SY}
\title{Get adult distribution (FXXX_SY) values from a biology prm file}
\usage{
get_param_FXXX_SY(bio.prm)
}
\arguments{
\item{bio.prm}{path to the biology prm file}
}
\value{
a data frame with the group, value and polygons, season, cohort
}
\description{
For each of the 4 seasons get the values (proportion of the population)
that are distributed among polygons.
}
