% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_move_temp.R
\name{get_param_move_temp}
\alias{get_param_move_temp}
\title{Get min_move_temp & max_move_temp values from a biology prm file}
\usage{
get_param_move_temp(bio.prm)
}
\arguments{
\item{bio.prm}{path to the biology prm file}
}
\value{
a data frame with the group, value and limit (max/min)
}
\description{
The values are the minimum and maximum temperatures that a group can tolerate
XXX_min_move_temp, XXX_max_move_temp
}
