% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_recruit_hdistrib.R
\name{get_param_recruit_hdistrib}
\alias{get_param_recruit_hdistrib}
\title{Get horizontal distribution recruitment values from a biology prm file}
\usage{
get_param_recruit_hdistrib(bio.prm)
}
\arguments{
\item{bio.prm}{path to the biology prm file}
}
\value{
a data frame with the group, value and polygon
}
\description{
The values are proportion of recruits and how they are distributed among
the polygons. XXX_recruit_hdistrib
}
