% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_recruit_vdistrib.R
\name{get_param_recruit_vdistrib}
\alias{get_param_recruit_vdistrib}
\title{Get vertical distribution recruitment values from a biology prm file}
\usage{
get_param_recruit_vdistrib(bio.prm)
}
\arguments{
\item{bio.prm}{path to the biology prm file}
}
\value{
a data frame with the group, value and layer
}
\description{
The values are proportion of recruits and how they are distributed among
the polygon layers. Layer 1 = surface, 4 = deepest. Note: Sediment layer is absent.
XXX_recruit_vdistrib
}
