% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_vert.R
\name{get_param_vert}
\alias{get_param_vert}
\title{Get adult distribution (vertDAY_XXX, vertNIGHT_XXX) values from a biology prm file}
\usage{
get_param_vert(bio.prm)
}
\arguments{
\item{bio.prm}{path to the biology prm file}
}
\value{
a data frame with the group, cohort, value and layer
}
\description{
Get the values (proportion of the population) that define the distribution among layers
of a boxes for each 12 hour period.
}
