% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_prm.R
\name{get_run_prm}
\alias{get_run_prm}
\title{Get parameters from the run prm file}
\usage{
get_run_prm(runprm, paramName = NULL)
}
\arguments{
\item{runprm}{path to the run prm file}

\item{paramName}{name of the parameter to extract}
}
\value{
list
\item{value}{value of parameter}
\item{unit}{units of parameter, if applicable}
}
\description{
Pull out the parameters from the run prm file. This will only return the values of
parameters that have a scalar value. For example, if paramName = init_scalar, only the length
of the vector will be returned.
}
